Move querylog into its own module
